added space condition for input and inputstring
